+++ docs/changelog.rst
@@ -39,6 +39,8 @@ Fixes:
 * Date fields are now written in the correct order (year-month-day), which
   eliminates an intermittent bug where the latter two fields would not get
   written to files. Thanks to :user:`jdetrey`. :bug:`1303` :bug:`1589`
+* :doc:`/plugins/replaygain`: Avoid a crash when the PyAudioTools backend
+  encounters an error. :bug:`1592`
 
 
 1.3.14 (August 2, 2015)